backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
celery>=5.3.0
redis>=4.5.0
# psycopg2-binary==2.9.10 --only-binary :all:  # Removed PostgreSQL
pyodbc==4.0.39  # Added SQL Server
mssql-django==1.5  # Added SQL Server backend (modern Django 4.x compatible)
factory-boy==3.3.1
pytest-django==4.7.0
pytest-cov==4.1.0
python-dotenv==1.0.1
django-storages==1.14.2
boto3==1.34.34
django-cors-headers==4.3.1
django-celery-beat>=2.5.0
gunicorn
daphne==4.0.0
django-filter
channels==4.0.0
channels-redis==4.1.0 
persiantools 
asgiref==3.7.2 
twisted==22.10.0
incremental==21.3.0
# LDAP Authentication (optional - will fallback to JSON if not available)
# python-ldap==3.4.3  # Uncomment if you can install it 